chore(ir): Refactor; restructure to a separate module
Will be useful in general, especially when working on #56
