+++ src/internals/ir/index.ts
@@ -1,0 +1,13 @@
+export {
+  Edge,
+  NodeKind,
+  Node,
+  FunctionKind,
+  CFG,
+  Contract,
+  getPredecessors,
+  getSuccessors,
+} from "./cfg";
+export { TactASTStore } from "./astStore";
+export { CompilationUnit } from "./ir";
+export * from "./types";
