+++ src/internals/ir/types.ts
@@ -1,0 +1,8 @@
+export type ProjectName = string;
+export type FunctionName = string;
+export type ContractName = string;
+
+export type EdgeIdx = number;
+export type NodeIdx = number;
+export type CFGIdx = number;
+export type ContractIdx = number;
