+++ test/testUtil.ts
@@ -1,4 +1,4 @@
-import { IdxGenerator } from "../src/internals/ir";
+import { IdxGenerator } from "../src/internals/ir/indices";
 import { __DANGER_resetNodeId } from "@tact-lang/compiler/dist/grammar/ast";
 import { expect } from "@jest/globals";
 import * as fs from "fs";
